refactor: update package structure that was broken during merging step
