+++ src/main/java/com/sw/yutnori/controller/GameSetupController.java
@@ -4,8 +4,7 @@ package com.sw.yutnori.controller;
 import com.sw.yutnori.logic.GameManager;
 import com.sw.yutnori.model.Board;
 import com.sw.yutnori.model.Player;
-import com.sw.yutnori.ui.swing.GameSetupDisplay;
-import com.sw.yutnori.controller.InGameController;
+import com.sw.yutnori.ui.display.GameSetupDisplay;
 
 import java.util.*;
 import java.util.function.Consumer;
