+++ src/main/java/com/sw/yutnori/logic/GameManager.java
@@ -3,7 +3,7 @@ package com.sw.yutnori.logic;
 
 import com.sw.yutnori.model.*;
 import com.sw.yutnori.model.enums.*;
-import com.sw.yutnori.ui.swing.GameSetupDisplay;
+import com.sw.yutnori.ui.display.GameSetupDisplay;
 
 import java.util.*;
