+++ src/main/java/com/sw/yutnori/ui/display/GameSetupDisplay.java
@@ -5,7 +5,7 @@
  * 
  * 
  */
-package com.sw.yutnori.ui.swing;
+package com.sw.yutnori.ui.display;
 import javax.swing.*;
 
 public interface GameSetupDisplay {
