+++ src/main/java/com/sw/yutnori/ui/display/ResultDisplay.java
@@ -1,4 +1,4 @@
-package com.sw.yutnori.ui.swing;
+package com.sw.yutnori.ui.display;
 
 public interface ResultDisplay {
