+++ src/main/java/com/sw/yutnori/ui/swing/panel/SwingYutBoardPanel.java
@@ -7,7 +7,6 @@
  */
 package com.sw.yutnori.ui.swing.panel;
 
-import com.sw.yutnori.board.*;
 import javax.swing.*;
 import java.awt.*;
 import java.util.List;
@@ -16,29 +15,30 @@ import java.awt.event.MouseEvent;
 import java.util.HashMap;
 import java.util.Map;
 
-import com.sw.yutnori.board.BoardModel;
-import com.sw.yutnori.common.LogicalPosition;
 import com.sw.yutnori.controller.InGameController;
 import com.sw.yutnori.logic.GameManager;
+import com.sw.yutnori.logic.util.ColorUtils;
+import com.sw.yutnori.model.Board;
+import com.sw.yutnori.model.LogicalPosition;
+import com.sw.yutnori.model.Node;
 import com.sw.yutnori.model.Piece;
 import com.sw.yutnori.model.Player;
-import com.sw.yutnori.ui.swing.PiecePositionDisplayManager;
-import com.sw.yutnori.util.ColorUtils;
 import com.sw.yutnori.model.enums.PieceState;
+import com.sw.yutnori.ui.swing.PiecePositionDisplayManager;
 
 public class SwingYutBoardPanel extends JPanel {
-    private final BoardModel boardModel;
+    private final Board board;
     private static final int BOARD_WIDTH = 1200;
     private static final int BOARD_HEIGHT = 1000;
-    private final Map<Long, JButton> pieceButtons = new HashMap<>();
+    private final Map<Long, JComponent> pieceButtons = new HashMap<>();
     private InGameController controller;
     private List<Piece> pieceList;
     private GameManager gameManager;
 
-    public SwingYutBoardPanel(BoardModel boardModel) {
-        this.boardModel = boardModel;
+    public SwingYutBoardPanel(Board board) {
+        this.board = board;
         setLayout(null);
-        setPreferredSize(new Dimension(boardModel.getWidth(), boardModel.getHeight()));
+        setPreferredSize(new Dimension(board.getWidth(), board.getHeight()));
     }
     public void setInGameController(InGameController controller) {
         this.controller = controller;
@@ -56,7 +56,7 @@ public class SwingYutBoardPanel extends JPanel {
         int offsetX = (panelW - BOARD_WIDTH) / 2;
         int offsetY = (panelH - BOARD_HEIGHT) / 2;
 
-        List<Node> boardNodes = boardModel.getNodes();
+        List<Node> boardNodes = board.getNodes();
         Graphics2D g2 = (Graphics2D) g;
         g2.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON);
 
@@ -115,7 +115,7 @@ public class SwingYutBoardPanel extends JPanel {
         }
     }
     private Rectangle getPieceBounds(Piece piece) {
-        Node node = boardModel.findNode(piece.getA(), piece.getB());
+        Node node = board.findNode(piece.getA(), piece.getB());
         if (node == null) return new Rectangle(); // or throw exception
         int nodeX = (int) node.getX();
         int nodeY = (int) node.getY();
@@ -136,7 +136,7 @@ public class SwingYutBoardPanel extends JPanel {
             pieceBtn.setBounds(x, y, 80, 40);
             pieceBtn.setBackground(Color.LIGHT_GRAY);
             pieceBtn.addActionListener(e -> {
-                highlightSelectedPiece(piece.getPieceId());
+                // highlightSelectedPiece(piece.getPieceId());
                 controller.setSelectedPieceId(piece.getPieceId());
             });
             pieceButtons.put(piece.getPieceId(), pieceBtn);
@@ -188,7 +188,7 @@ public class SwingYutBoardPanel extends JPanel {
             pieceBtn.setBounds(x, y, 80, 40);
             pieceBtn.setBackground(Color.LIGHT_GRAY);
             pieceBtn.addActionListener(e -> {
-                highlightSelectedPiece(pieceId);
+                // highlightSelectedPiece(pieceId);
                 controller.setSelectedPieceId(pieceId);
             });
             pieceButtons.put(pieceId, pieceBtn);
@@ -208,7 +208,7 @@ public class SwingYutBoardPanel extends JPanel {
                 currentPosition = new LogicalPosition(clickedPos.getA(), clickedPos.getB());
                 controlPanel.enableYutSelection(); // 윷 선택 UI 열기
                 // 논리 좌표 기반 말 위치 표시
-                PiecePositionDisplayManager markerManager = new PiecePositionDisplayManager(boardModel, this, gameManager);
+                PiecePositionDisplayManager markerManager = new PiecePositionDisplayManager(board, this, gameManager);
                 markerManager.showLogicalPosition(currentPosition, selectedPieceId);
 
             }
@@ -229,35 +229,36 @@ public class SwingYutBoardPanel extends JPanel {
             Color color = ColorUtils.parseColor(player.getColor());
             for (Piece piece : player.getPieces()) {
                 if (piece.getState() == PieceState.ON_BOARD && !piece.isFinished()) {
-                    Node node = boardModel.findNode(piece.getA(), piece.getB());
+                    Node node = board.findNode(piece.getA(), piece.getB());
                     if (node == null) continue;
-                    // 노드 중심과 버튼 중앙이 일치하도록 배치
                     int x = (int) node.getX() + offsetX - (pieceSize / 2);
                     int y = (int) node.getY() + offsetY - (pieceSize / 2);
-                    // 말 번호는 양측 모두 1부터 n(2<=n<=5) 순서대로 표시
                     int displayNum = player.getPieces().indexOf(piece) + 1;
-                    JButton btn = new JButton(String.valueOf(displayNum));
-                    btn.setBounds(x, y, pieceSize, pieceSize);
-                    btn.setBackground(color);
-                    btn.setOpaque(true);
-                    btn.setBorderPainted(false);
-                    btn.setMargin(new Insets(0, 0, 0, 0));
-                    pieceButtons.put(piece.getPieceId(), btn);
-                    add(btn);
+                    // JButton을 사용하니 색상이 제대로 출력되지 않는 문제가 있었음. -> JLabel로 변경
+                    JLabel pieceLabel = new JLabel(String.valueOf(displayNum), SwingConstants.CENTER);
+                    pieceLabel.setBounds(x, y, pieceSize, pieceSize);
+                    pieceLabel.setOpaque(true);
+                    pieceLabel.setBackground(color);
+                    pieceLabel.setForeground(Color.BLACK);
+                    pieceLabel.setFont(new Font("맑은 고딕", Font.PLAIN, 16));
+                    pieceButtons.put(piece.getPieceId(), pieceLabel);
+                    add(pieceLabel);
                 }
             }
         }
         revalidate();
         repaint();
     }
-    // 선택된 말 강조 표시
-    private void highlightSelectedPiece(Long selectedId) {
-        for (Map.Entry<Long, JButton> entry : pieceButtons.entrySet()) {
-            if (entry.getKey().equals(selectedId)) {
-                entry.getValue().setBackground(Color.ORANGE);
-            } else {
-                entry.getValue().setBackground(Color.LIGHT_GRAY);
-            }
-        }
-    }
+    // // 선택된 말 강조 표시 (JLabel용) -> 굳이 필요한지 잘 모르겠음
+    // public void highlightSelectedPiece(Long selectedId) {
+    //     for (Map.Entry<Long, JComponent> entry : pieceButtons.entrySet()) {
+    //         if (entry.getValue() instanceof JLabel label) {
+    //             if (entry.getKey().equals(selectedId)) {
+    //                 label.setBorder(BorderFactory.createLineBorder(Color.ORANGE, 3));
+    //             } else {
+    //                 label.setBorder(BorderFactory.createLineBorder(Color.DARK_GRAY, 2));
+    //             }
+    //         }
+    //     }
+    // }
 }
